feat: entity

+++ src/main/java/com/nhnacademy/springbootjpaassignment/entity/Cart.java
@@ -11,7 +11,7 @@ import lombok.*;
 public class Cart {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer cartId;
+    private Long cartId;
 
     @Column(name = "user_id", nullable = false)
     private String userId;
