+++ src/main/java/com/nhnacademy/springbootjpaassignment/entity/ProductCategory.java
@@ -13,7 +13,7 @@ public class ProductCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "pc_id")
-    private Integer pcId;
+    private Long pcId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "product_id")
